ai_customer_support/requirements.txt
flask
sentence-transformers
qdrant-client
numpy
python-dotenv
groq
